PB-17b: renamed adminpage to admin section of app
