+++ application/next-frontend/src/Admin/components/CardComponent.tsx
