+++ application/next-frontend/src/Admin/scenarios/Events/components/CreateEventCard.tsx
