+++ application/next-frontend/src/Admin/scenarios/Navbar.tsx
